+++ docker-compose.yml
@@ -28,7 +28,7 @@ services:
     environment:
       - APP_SETTINGS=project.config.DevelopmentConfig
     env_file:
-      - "~/.team-matcha-secrets"
+      - "~/.team-matcha-secrets.env"
     depends_on:
       - postgres-db
