+++ server/project/test/user_test.py
@@ -2,14 +2,6 @@ import json
 from project import db
 from project.models.user import User, add_user
 from project.test.test_base import TestBase
-import uuid
-
-
-def add_user(name='kenny', email='test@email.com'):
-    user = User(public_id=uuid.uuid4(), name=name, email=email)
-    db.session.add(user)
-    db.session.commit()
-    return user
 
 
 class UserModelTest(TestBase):
